src/requirements.txt
requests
google-cloud-storage
functions-framework
pymupdf
flask
python-dotenv
